requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
psycopg2-binary==2.8.5
python-jose==3.2.0
